requirements.txt
django==3.1
markdown==3.2.2
gunicorn==20.0.4
Pygments==2.6.1
django-anymail[mailgun]==7.2.1
pillow==7.2.0
